Add sliding window method
This adds the ability to generate chips using a sliding window
for object detection, similar to how it works for classification.
We need this to ensure that the chips used for making predictions
are exactly the same as the ones used to train the model, which is
needed to overfit the model very quickly.