Prototype of workflow library
Previously, we compiled a chain workflow config file into a set of command
config files. This was difficult for beginners because it requires
understanding how the compilation process works, which also happens to be
poorly documented. It was also inflexible and did not encapsulate knowledge of good
defaults. The new way of generating a chain workflow is to use a library of
utilities for programmatically creating configs, which attempts to fix the above
limitations.

* rastervision.workflows.config_utils contains generic functions for creating
various protobuf configs
* rastervision.workflows.chain_utils contains more opinionated classes for
creating chain workflows
* rastervision.workflows.sample is a script that generates a workflow using the
utils. If you change the base_uri field and then run it, the script should
generate a chain workflow json file, which functions as the artifact/manifest
files we've discussed.

Some details are missing, but this gives an idea of the direction we are moving
in.